move bulletin constant to test helpers
