+++ bin/orbis-node/src/dkg/tests/dkg.rs
@@ -5,6 +5,7 @@ use crate::dkg::{
     session_state::{CreateSessionOutcome, DkgMessageType, DkgPhase, SessionStateManager},
 };
 use crate::helpers::helpers::extract_node_part;
+use crate::helpers::test_helpers::BULLETIN_RING_NAMESPACE;
 use crate::helpers::test_helpers::{
     cleanup_db, create_authenticated_request, create_test_app_state, create_test_app_state_default,
     get_test_ring_post, setup_three_node_network, test_db_path, TestKeyPair,
@@ -35,18 +36,13 @@ async fn test_start_dkg_empty_participants() {
         threshold: 0,
         peer_ids: peer_ids.clone(),
         pss_interval: None,
-        namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+        namespace: BULLETIN_RING_NAMESPACE.to_string(),
     };
 
     // Create authenticated request
     let test_keys = TestKeyPair::new();
     let token = test_keys
-        .create_dkg_jwt(
-            0,
-            &peer_ids,
-            None,
-            crate::constants::BULLETIN_RING_NAMESPACE,
-        )
+        .create_dkg_jwt(0, &peer_ids, None, BULLETIN_RING_NAMESPACE)
         .expect("Failed to create JWT");
     let tonic_request = create_authenticated_request(request, &token).unwrap();
 
@@ -87,18 +83,13 @@ async fn test_three_nodes_connect() {
         threshold: 2,
         peer_ids: peer_ids.clone(),
         pss_interval: None,
-        namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+        namespace: BULLETIN_RING_NAMESPACE.to_string(),
     };
 
     // Create authenticated request
     let test_keys = TestKeyPair::new();
     let token = test_keys
-        .create_dkg_jwt(
-            2,
-            &peer_ids,
-            None,
-            crate::constants::BULLETIN_RING_NAMESPACE,
-        )
+        .create_dkg_jwt(2, &peer_ids, None, BULLETIN_RING_NAMESPACE)
         .expect("Failed to create JWT");
 
     println!("Alice sending StartDkgRequest with peer IDs...");
@@ -157,18 +148,13 @@ async fn test_start_dkg_fails_on_connection_failure() {
         threshold: 2,
         peer_ids: peer_ids.clone(),
         pss_interval: None,
-        namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+        namespace: BULLETIN_RING_NAMESPACE.to_string(),
     };
 
     // Create authenticated request (even with invalid peer_ids, JWT should match request)
     let test_keys = TestKeyPair::new();
     let token = test_keys
-        .create_dkg_jwt(
-            2,
-            &peer_ids,
-            None,
-            crate::constants::BULLETIN_RING_NAMESPACE,
-        )
+        .create_dkg_jwt(2, &peer_ids, None, BULLETIN_RING_NAMESPACE)
         .expect("Failed to create JWT");
 
     println!("Alice sending StartDkgRequest with invalid peer IDs...");
@@ -232,18 +218,13 @@ async fn test_start_dkg_succeeds_on_all_connections() {
         threshold: 2,
         peer_ids: peer_ids.clone(),
         pss_interval: None,
-        namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+        namespace: BULLETIN_RING_NAMESPACE.to_string(),
     };
 
     // Create authenticated request
     let test_keys = TestKeyPair::new();
     let token = test_keys
-        .create_dkg_jwt(
-            2,
-            &peer_ids,
-            None,
-            crate::constants::BULLETIN_RING_NAMESPACE,
-        )
+        .create_dkg_jwt(2, &peer_ids, None, BULLETIN_RING_NAMESPACE)
         .expect("Failed to create JWT");
 
     println!("Alice sending StartDkgRequest with valid peer IDs...");
@@ -380,7 +361,7 @@ async fn test_start_dkg_fails_missing_auth_header() {
         threshold: 2,
         peer_ids,
         pss_interval: None,
-        namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+        namespace: BULLETIN_RING_NAMESPACE.to_string(),
     };
 
     // Create request WITHOUT authentication header
@@ -420,7 +401,7 @@ async fn test_start_dkg_fails_malformed_jwt() {
         threshold: 2,
         peer_ids,
         pss_interval: None,
-        namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+        namespace: BULLETIN_RING_NAMESPACE.to_string(),
     };
 
     // Create request with malformed JWT (not a valid JWT structure)
@@ -454,12 +435,7 @@ async fn test_start_dkg_fails_wrong_signature() {
     // Create a valid JWT with key_pair_1
     let key_pair_1 = TestKeyPair::new();
     let valid_token = key_pair_1
-        .create_dkg_jwt(
-            2,
-            &peer_ids,
-            None,
-            crate::constants::BULLETIN_RING_NAMESPACE,
-        )
+        .create_dkg_jwt(2, &peer_ids, None, BULLETIN_RING_NAMESPACE)
         .expect("Failed to create JWT");
 
     // Tamper with the signature by changing a character
@@ -480,7 +456,7 @@ async fn test_start_dkg_fails_wrong_signature() {
         threshold: 2,
         peer_ids,
         pss_interval: None,
-        namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+        namespace: BULLETIN_RING_NAMESPACE.to_string(),
     };
 
     let tonic_request = create_authenticated_request(request, &tampered_token).unwrap();
@@ -532,7 +508,7 @@ async fn test_dkg_session_init_fails_with_invalid_jwt() {
         token_string: "not-a-valid-jwt-token".to_string(), // Invalid JWT
         kind: SessionKind::Fresh,
         pss_interval: None,
-        namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+        namespace: BULLETIN_RING_NAMESPACE.to_string(),
     };
 
     // Try to handle the message - should fail due to invalid JWT
@@ -584,12 +560,7 @@ async fn test_dkg_session_init_fails_with_mismatched_claims() {
 
     // Create JWT with threshold=3, but SessionInit will have threshold=2
     let mismatched_token = test_keys
-        .create_dkg_jwt(
-            3,
-            &peer_ids,
-            None,
-            crate::constants::BULLETIN_RING_NAMESPACE,
-        ) // Wrong threshold!
+        .create_dkg_jwt(3, &peer_ids, None, BULLETIN_RING_NAMESPACE) // Wrong threshold!
         .expect("Failed to create JWT");
 
     // Create a SessionInit message with threshold=2 (doesn't match JWT's threshold=3)
@@ -606,7 +577,7 @@ async fn test_dkg_session_init_fails_with_mismatched_claims() {
         token_string: mismatched_token,
         kind: SessionKind::Fresh,
         pss_interval: None,
-        namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+        namespace: BULLETIN_RING_NAMESPACE.to_string(),
     };
 
     // Try to handle the message - should fail due to claim mismatch
@@ -649,12 +620,7 @@ async fn test_dkg_session_init_fails_with_wrong_peer_ids() {
     ];
 
     let mismatched_token = test_keys
-        .create_dkg_jwt(
-            2,
-            &jwt_peer_ids,
-            None,
-            crate::constants::BULLETIN_RING_NAMESPACE,
-        )
+        .create_dkg_jwt(2, &jwt_peer_ids, None, BULLETIN_RING_NAMESPACE)
         .expect("Failed to create JWT");
 
     // SessionInit has different peer_ids than the JWT
@@ -677,7 +643,7 @@ async fn test_dkg_session_init_fails_with_wrong_peer_ids() {
         token_string: mismatched_token,
         kind: SessionKind::Fresh,
         pss_interval: None,
-        namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+        namespace: BULLETIN_RING_NAMESPACE.to_string(),
     };
 
     // Try to handle the message - should fail due to peer_ids mismatch
