+++ bin/orbis-node/src/dkg/tests/reshare.rs
@@ -1,4 +1,3 @@
-use crate::constants::BULLETIN_RING_NAMESPACE;
 use crate::dkg::{
     coordinator::DkgCoordinator,
     helpers::derive_reshare_session_id,
@@ -6,6 +5,7 @@ use crate::dkg::{
 };
 use crate::helpers::create_routers::create_router_with_all_handlers;
 use crate::helpers::helpers::extract_node_part;
+use crate::helpers::test_helpers::BULLETIN_RING_NAMESPACE;
 use crate::helpers::test_helpers::{
     cleanup_db, create_authenticated_request, create_test_app_state_default,
     create_test_app_state_with_bulletin, get_test_ring_post, setup_three_node_network,
@@ -66,7 +66,7 @@ fn reshare_session_init(
             bulletin_post_id: String::new(),
         },
         pss_interval: None,
-        namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+        namespace: BULLETIN_RING_NAMESPACE.to_string(),
     }
 }
 
@@ -158,7 +158,6 @@ async fn test_reshare_session_init_rejects_mismatched_bulletin_ring_pk() {
     let db_path = test_db_path(db_name);
     let app_state = Arc::new(create_test_app_state_default(db_name).await);
 
-    use crate::constants::BULLETIN_RING_NAMESPACE;
     use local_storage::r#trait::LocalStorageKeys;
 
     let sender_hex = "aabbccdd";
@@ -193,7 +192,7 @@ async fn test_reshare_session_init_rejects_mismatched_bulletin_ring_pk() {
     ring_index.push(RingIndexEntry {
         ring_pk_str: session_ring_pk.to_string(),
         bulletin_post_id: post_id,
-        bulletin_namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+        bulletin_namespace: BULLETIN_RING_NAMESPACE.to_string(),
     });
     app_state
         .local_storage
@@ -468,7 +467,6 @@ async fn write_ring_with_announced_reshare(
     announced_new_peer_ids: Option<Vec<String>>,
     announced_new_threshold: Option<u32>,
 ) {
-    use crate::constants::BULLETIN_RING_NAMESPACE;
     use crate::ring_state::RingIndexEntry;
     use local_storage::r#trait::LocalStorageKeys;
 
@@ -501,7 +499,7 @@ async fn write_ring_with_announced_reshare(
     ring_index.push(RingIndexEntry {
         ring_pk_str: ring_pk.to_string(),
         bulletin_post_id: post_id,
-        bulletin_namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+        bulletin_namespace: BULLETIN_RING_NAMESPACE.to_string(),
     });
     app_state
         .local_storage
@@ -750,7 +748,6 @@ async fn post_ring_for_validation(
     new_peer_ids: Option<Vec<String>>,
     new_threshold: Option<u32>,
 ) {
-    use crate::constants::BULLETIN_RING_NAMESPACE;
     use crate::ring_state::RingIndexEntry;
     use local_storage::r#trait::LocalStorageKeys;
 
@@ -783,7 +780,7 @@ async fn post_ring_for_validation(
     ring_index.push(RingIndexEntry {
         ring_pk_str: ring_pk.to_string(),
         bulletin_post_id: post_id,
-        bulletin_namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+        bulletin_namespace: BULLETIN_RING_NAMESPACE.to_string(),
     });
     app_state
         .local_storage
@@ -824,7 +821,7 @@ async fn test_validate_reshare_accepts_new_peer_ids_fallback_to_current() {
         &[sender_hex.to_string()], // matches fallback = peer_ids
         1,
         "",
-        crate::constants::BULLETIN_RING_NAMESPACE,
+        BULLETIN_RING_NAMESPACE,
         &app_state.local_storage,
         &app_state.bulletin,
     )
@@ -868,7 +865,7 @@ async fn test_validate_reshare_accepts_new_threshold_fallback_to_current() {
         &[new_peer.to_string()],
         1, // matches fallback = current threshold
         "",
-        crate::constants::BULLETIN_RING_NAMESPACE,
+        BULLETIN_RING_NAMESPACE,
         &app_state.local_storage,
         &app_state.bulletin,
     )
@@ -911,7 +908,7 @@ async fn test_validate_reshare_rejects_when_peers_differ_from_fallback() {
         &["00112233".to_string()], // differs from fallback = ["aabbccdd"]
         1,
         "",
-        crate::constants::BULLETIN_RING_NAMESPACE,
+        BULLETIN_RING_NAMESPACE,
         &app_state.local_storage,
         &app_state.bulletin,
     )
@@ -956,7 +953,7 @@ async fn test_validate_reshare_rejects_when_threshold_differs_from_fallback() {
         &[new_peer_1.to_string(), new_peer_2.to_string()],
         1, // differs from fallback = 2
         "",
-        crate::constants::BULLETIN_RING_NAMESPACE,
+        BULLETIN_RING_NAMESPACE,
         &app_state.local_storage,
         &app_state.bulletin,
     )
@@ -1075,16 +1072,12 @@ async fn post_reshare_announcement(
     let bytes = serde_json::to_vec(&payload).unwrap();
 
     bulletin
-        .post(
-            crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
-            bytes.clone(),
-            None,
-        )
+        .post(BULLETIN_RING_NAMESPACE.to_string(), bytes.clone(), None)
         .await
         .expect("post reshare announcement to bulletin");
 
     let new_post_id = bulletin
-        .get_post_id(crate::constants::BULLETIN_RING_NAMESPACE, &bytes)
+        .get_post_id(BULLETIN_RING_NAMESPACE, &bytes)
         .expect("get reshare announcement post_id");
 
     // Point every old-committee node's RingIndex entry at the new post.
@@ -1176,7 +1169,7 @@ async fn run_reshare_ceremony(
             bulletin_post_id: bulletin_post_id.to_string(),
         },
         pss_interval: None,
-        namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+        namespace: BULLETIN_RING_NAMESPACE.to_string(),
     };
 
     // Process own SessionInit — sets up session state and reshare_params.
@@ -1322,12 +1315,7 @@ async fn test_reshare_lower_threshold() {
     let alice_service = DkgServiceImpl::<DkgImpl>::new(network.alice.app_state.clone());
     let test_keys = TestKeyPair::new();
     let token = test_keys
-        .create_dkg_jwt(
-            2,
-            &peer_ids,
-            None,
-            crate::constants::BULLETIN_RING_NAMESPACE,
-        )
+        .create_dkg_jwt(2, &peer_ids, None, BULLETIN_RING_NAMESPACE)
         .expect("JWT");
     alice_service
         .start_dkg(
@@ -1336,7 +1324,7 @@ async fn test_reshare_lower_threshold() {
                     threshold: 2,
                     peer_ids: peer_ids.clone(),
                     pss_interval: None,
-                    namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+                    namespace: BULLETIN_RING_NAMESPACE.to_string(),
                 },
                 &token,
             )
@@ -1434,12 +1422,7 @@ async fn test_reshare_one_member_rotated() {
     let alice_service = DkgServiceImpl::<DkgImpl>::new(network.alice.app_state.clone());
     let test_keys = TestKeyPair::new();
     let token = test_keys
-        .create_dkg_jwt(
-            2,
-            &peer_ids,
-            None,
-            crate::constants::BULLETIN_RING_NAMESPACE,
-        )
+        .create_dkg_jwt(2, &peer_ids, None, BULLETIN_RING_NAMESPACE)
         .expect("JWT");
     alice_service
         .start_dkg(
@@ -1448,7 +1431,7 @@ async fn test_reshare_one_member_rotated() {
                     threshold: 2,
                     peer_ids: peer_ids.clone(),
                     pss_interval: None,
-                    namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+                    namespace: BULLETIN_RING_NAMESPACE.to_string(),
                 },
                 &token,
             )
@@ -1549,12 +1532,7 @@ async fn test_reshare_one_old_dealer_offline_completes() {
     let alice_service = DkgServiceImpl::<DkgImpl>::new(network.alice.app_state.clone());
     let test_keys = TestKeyPair::new();
     let token = test_keys
-        .create_dkg_jwt(
-            2,
-            &peer_ids,
-            None,
-            crate::constants::BULLETIN_RING_NAMESPACE,
-        )
+        .create_dkg_jwt(2, &peer_ids, None, BULLETIN_RING_NAMESPACE)
         .expect("JWT");
     alice_service
         .start_dkg(
@@ -1563,7 +1541,7 @@ async fn test_reshare_one_old_dealer_offline_completes() {
                     threshold: 2,
                     peer_ids: peer_ids.clone(),
                     pss_interval: None,
-                    namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+                    namespace: BULLETIN_RING_NAMESPACE.to_string(),
                 },
                 &token,
             )
@@ -1671,12 +1649,7 @@ async fn test_reshare_expand_committee() {
     let alice_service = DkgServiceImpl::<DkgImpl>::new(network.alice.app_state.clone());
     let test_keys = TestKeyPair::new();
     let token = test_keys
-        .create_dkg_jwt(
-            2,
-            &peer_ids,
-            None,
-            crate::constants::BULLETIN_RING_NAMESPACE,
-        )
+        .create_dkg_jwt(2, &peer_ids, None, BULLETIN_RING_NAMESPACE)
         .expect("JWT");
     alice_service
         .start_dkg(
@@ -1685,7 +1658,7 @@ async fn test_reshare_expand_committee() {
                     threshold: 2,
                     peer_ids: peer_ids.clone(),
                     pss_interval: None,
-                    namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+                    namespace: BULLETIN_RING_NAMESPACE.to_string(),
                 },
                 &token,
             )
@@ -1793,12 +1766,7 @@ async fn test_reshare_shrink_committee() {
     let alice_service = DkgServiceImpl::<DkgImpl>::new(network.alice.app_state.clone());
     let test_keys = TestKeyPair::new();
     let token = test_keys
-        .create_dkg_jwt(
-            2,
-            &peer_ids,
-            None,
-            crate::constants::BULLETIN_RING_NAMESPACE,
-        )
+        .create_dkg_jwt(2, &peer_ids, None, BULLETIN_RING_NAMESPACE)
         .expect("JWT");
     alice_service
         .start_dkg(
@@ -1807,7 +1775,7 @@ async fn test_reshare_shrink_committee() {
                     threshold: 2,
                     peer_ids: peer_ids.clone(),
                     pss_interval: None,
-                    namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+                    namespace: BULLETIN_RING_NAMESPACE.to_string(),
                 },
                 &token,
             )
@@ -1904,12 +1872,7 @@ async fn test_reshare_full_rotation() {
     let alice_service = DkgServiceImpl::<DkgImpl>::new(network.alice.app_state.clone());
     let test_keys = TestKeyPair::new();
     let token = test_keys
-        .create_dkg_jwt(
-            2,
-            &peer_ids,
-            None,
-            crate::constants::BULLETIN_RING_NAMESPACE,
-        )
+        .create_dkg_jwt(2, &peer_ids, None, BULLETIN_RING_NAMESPACE)
         .expect("JWT");
     alice_service
         .start_dkg(
@@ -1918,7 +1881,7 @@ async fn test_reshare_full_rotation() {
                     threshold: 2,
                     peer_ids: peer_ids.clone(),
                     pss_interval: None,
-                    namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+                    namespace: BULLETIN_RING_NAMESPACE.to_string(),
                 },
                 &token,
             )
