+++ bin/orbis-node/src/pre/tests.rs
@@ -5,7 +5,7 @@
 
 use crate::helpers::test_helpers::{
     cleanup_db, create_authenticated_request, create_test_app_state_default, get_test_ring_post,
-    setup_three_node_network_with_pre, test_db_path, TestKeyPair,
+    setup_three_node_network_with_pre, test_db_path, TestKeyPair, BULLETIN_RING_NAMESPACE,
 };
 use crate::pre::coordinator::{PreCoordinator, PreResponse};
 use crate::pre::service::PreServiceImpl;
@@ -127,18 +127,13 @@ async fn test_dkg_then_pre_end_to_end() {
         threshold: 2,
         peer_ids: peer_ids.clone(),
         pss_interval: None,
-        namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+        namespace: BULLETIN_RING_NAMESPACE.to_string(),
     };
 
     // Create authenticated request
     let test_keys = TestKeyPair::new();
     let token = test_keys
-        .create_dkg_jwt(
-            2,
-            &peer_ids,
-            None,
-            crate::constants::BULLETIN_RING_NAMESPACE,
-        )
+        .create_dkg_jwt(2, &peer_ids, None, BULLETIN_RING_NAMESPACE)
         .expect("Failed to create JWT");
 
     println!("Node1 sending StartDkgRequest...");
@@ -392,18 +387,13 @@ async fn test_pre_with_large_secret() {
         threshold: 2,
         peer_ids: peer_ids.clone(),
         pss_interval: None,
-        namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+        namespace: BULLETIN_RING_NAMESPACE.to_string(),
     };
 
     // Create authenticated request
     let test_keys = TestKeyPair::new();
     let token = test_keys
-        .create_dkg_jwt(
-            2,
-            &peer_ids,
-            None,
-            crate::constants::BULLETIN_RING_NAMESPACE,
-        )
+        .create_dkg_jwt(2, &peer_ids, None, BULLETIN_RING_NAMESPACE)
         .expect("Failed to create JWT");
 
     let result = node1_service
@@ -536,18 +526,13 @@ async fn test_pre_fails_with_wrong_key() {
         threshold: 2,
         peer_ids: peer_ids.clone(),
         pss_interval: None,
-        namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+        namespace: BULLETIN_RING_NAMESPACE.to_string(),
     };
 
     // Create authenticated request
     let test_keys = TestKeyPair::new();
     let token = test_keys
-        .create_dkg_jwt(
-            2,
-            &peer_ids,
-            None,
-            crate::constants::BULLETIN_RING_NAMESPACE,
-        )
+        .create_dkg_jwt(2, &peer_ids, None, BULLETIN_RING_NAMESPACE)
         .expect("Failed to create JWT");
 
     let result = node1_service
@@ -682,18 +667,13 @@ async fn test_pre_fails_with_invalid_jwt_token() {
         threshold: 2,
         peer_ids: peer_ids.clone(),
         pss_interval: None,
-        namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+        namespace: BULLETIN_RING_NAMESPACE.to_string(),
     };
 
     // Create authenticated request
     let test_keys = TestKeyPair::new();
     let token = test_keys
-        .create_dkg_jwt(
-            2,
-            &peer_ids,
-            None,
-            crate::constants::BULLETIN_RING_NAMESPACE,
-        )
+        .create_dkg_jwt(2, &peer_ids, None, BULLETIN_RING_NAMESPACE)
         .expect("Failed to create JWT");
 
     let result = node1_service
@@ -836,17 +816,12 @@ async fn test_pre_fails_with_mismatched_jwt_claims() {
         threshold: 2,
         peer_ids: peer_ids.clone(),
         pss_interval: None,
-        namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+        namespace: BULLETIN_RING_NAMESPACE.to_string(),
     };
 
     let test_keys = TestKeyPair::new();
     let token = test_keys
-        .create_dkg_jwt(
-            2,
-            &peer_ids,
-            None,
-            crate::constants::BULLETIN_RING_NAMESPACE,
-        )
+        .create_dkg_jwt(2, &peer_ids, None, BULLETIN_RING_NAMESPACE)
         .expect("Failed to create JWT");
 
     let result = node1_service
@@ -984,7 +959,7 @@ async fn test_start_pre_fails_missing_auth_header() {
 
     let request = StartPreRequest {
         rdr_pk: b"def456".to_vec(),
-        namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+        namespace: BULLETIN_RING_NAMESPACE.to_string(),
         object_id: "".to_string(),
         derivation: None,
         salt: None,
@@ -1026,7 +1001,7 @@ async fn test_start_pre_fails_malformed_jwt() {
 
     let request = StartPreRequest {
         rdr_pk: b"def456".to_vec(),
-        namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+        namespace: BULLETIN_RING_NAMESPACE.to_string(),
         object_id: "".to_string(),
         derivation: None,
         salt: None,
@@ -1085,7 +1060,7 @@ async fn test_start_pre_fails_wrong_signature() {
 
     let request = StartPreRequest {
         rdr_pk: b"def456".to_vec(),
-        namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+        namespace: BULLETIN_RING_NAMESPACE.to_string(),
         object_id: "".to_string(),
         derivation: None,
         salt: None,
@@ -1137,17 +1112,12 @@ async fn test_pre_fails_with_wrong_derivation() {
         threshold: 2,
         peer_ids: peer_ids.clone(),
         pss_interval: None,
-        namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+        namespace: BULLETIN_RING_NAMESPACE.to_string(),
     };
 
     let test_keys = TestKeyPair::new();
     let token = test_keys
-        .create_dkg_jwt(
-            2,
-            &peer_ids,
-            None,
-            crate::constants::BULLETIN_RING_NAMESPACE,
-        )
+        .create_dkg_jwt(2, &peer_ids, None, BULLETIN_RING_NAMESPACE)
         .expect("Failed to create JWT");
 
     let result = node1_service
@@ -1325,17 +1295,12 @@ async fn test_pre_fails_with_bad_proof() {
         threshold: 2,
         peer_ids: peer_ids.clone(),
         pss_interval: None,
-        namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+        namespace: BULLETIN_RING_NAMESPACE.to_string(),
     };
 
     let test_keys = TestKeyPair::new();
     let token = test_keys
-        .create_dkg_jwt(
-            2,
-            &peer_ids,
-            None,
-            crate::constants::BULLETIN_RING_NAMESPACE,
-        )
+        .create_dkg_jwt(2, &peer_ids, None, BULLETIN_RING_NAMESPACE)
         .expect("Failed to create JWT");
 
     let result = node1_service
