+++ bin/orbis-node/src/pss/tests.rs
@@ -1,6 +1,6 @@
-use crate::constants::BULLETIN_RING_NAMESPACE;
 use crate::dkg::error::DkgError;
 use crate::helpers::helpers::extract_node_part;
+use crate::helpers::test_helpers::BULLETIN_RING_NAMESPACE;
 use crate::helpers::test_helpers::{cleanup_db, create_test_app_state_with_bulletin, test_db_path};
 use crate::ring_state::RingIndexEntry;
 use bulletin::{
@@ -56,7 +56,7 @@ async fn make_state_with_ring(
     let entry = RingIndexEntry {
         ring_pk_str: ring_payload.ring_pk.clone(),
         bulletin_post_id: post_id,
-        bulletin_namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+        bulletin_namespace: BULLETIN_RING_NAMESPACE.to_string(),
     };
     let index_bytes = serde_json::to_vec(&vec![&entry]).expect("serialize RingIndex");
     app_state
@@ -150,7 +150,7 @@ async fn test_refresh_all_rings_bulletin_miss_does_not_propagate() {
     let ring_index = vec![RingIndexEntry {
         ring_pk_str: "nonexistent_ring".to_string(),
         bulletin_post_id: "nonexistent_ring_id".to_string(),
-        bulletin_namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+        bulletin_namespace: BULLETIN_RING_NAMESPACE.to_string(),
     }];
     let index_bytes = serde_json::to_vec(&ring_index).expect("serialize ring index");
     app_state
@@ -299,7 +299,7 @@ async fn test_refresh_ring_bad_bulletin_payload() {
     let entry = RingIndexEntry {
         ring_pk_str: ring_pk_str.to_string(),
         bulletin_post_id: post_id,
-        bulletin_namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+        bulletin_namespace: BULLETIN_RING_NAMESPACE.to_string(),
     };
     app_state
         .local_storage
@@ -354,7 +354,7 @@ async fn test_refresh_ring_rejects_bulletin_ring_pk_mismatch() {
     let entry = RingIndexEntry {
         ring_pk_str: "expected_ring_pk".to_string(),
         bulletin_post_id: post_id,
-        bulletin_namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+        bulletin_namespace: BULLETIN_RING_NAMESPACE.to_string(),
     };
     app_state
         .local_storage
@@ -418,7 +418,7 @@ async fn post_ring_and_seed_index(
     let entry = RingIndexEntry {
         ring_pk_str: ring_payload.ring_pk.clone(),
         bulletin_post_id: post_id,
-        bulletin_namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+        bulletin_namespace: BULLETIN_RING_NAMESPACE.to_string(),
     };
     app_state
         .local_storage
@@ -542,7 +542,7 @@ async fn test_refresh_ring_missing_from_bulletin() {
     let entry = RingIndexEntry {
         ring_pk_str: "ghost_ring".to_string(),
         bulletin_post_id: "ghost_ring".to_string(),
-        bulletin_namespace: crate::constants::BULLETIN_RING_NAMESPACE.to_string(),
+        bulletin_namespace: BULLETIN_RING_NAMESPACE.to_string(),
     };
     let result = super::pss_ring(&Arc::new(app_state), &entry).await;
     assert!(
